suppression de warning
git-svn-id: https://ctp-gesconf.canaltp.fr/svn/NAViTiACpp/trunk@196 a734f7ea-df10-47d3-a172-9f42111c9061

+++ source/utils/encoding_converter.cpp
@@ -16,7 +16,7 @@ std::string EncodingConverter::convert(std::string& str){
     size_t output_left = buffer_size;
     size_t input_left = str.size();
     size_t result = iconv(iconv_handler, &working_input, &input_left, &working_output, &output_left);
-    if(result == -1){
+    if(result == (size_t)-1){
         throw std::string("iconv fail: " + errno);
     }
     return std::string(iconv_output_buffer);
